requirements.txt
os_ken==3.1.1
ovs==3.7.0
ovsdbapp==2.14.1
prometheus_client==0.21.1
python-dotenv==1.2.2
